requirements.txt
annotated-types==0.7.0
anyio==4.9.0
atproto==0.0.61
certifi==2025.7.14
cffi==1.17.1
discord.py==2.3.2
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.5
dnspython==2.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
letta-client==0.1.235
libipld==3.1.1
markdown-it-py==3.0.0
mdurl==0.1.2
oauthlib==3.3.1
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
pygments==2.19.2
python-dotenv==1.1.1
pyyaml==6.0.2
requests==2.32.4
requests-oauthlib==2.0.0
rich==14.1.0
sniffio==1.3.1
typing-extensions==4.14.1
typing-inspection==0.4.1
urllib3==2.5.0
websockets==13.1
